requirements.txt
openai
python-dotenv
pandas
openpyxl
mysql-connector-python
